requirements.txt
twitchio
python-dotenv
openai
gtts
playsound
